Relocate header files shared by all platforms under Common.
