+++ src/Platforms/CUDA/CUDADeviceManager.cpp
@@ -15,7 +15,7 @@
 #include <stdexcept>
 #include "OutputManager.h"
 #include "CUDAruntime.hpp" // Positioned here to avoid conflict between CUDA and GCC >= 12 headers. https://github.com/QMCPACK/qmcpack/pull/4814
-#include "determineDefaultDeviceNum.h"
+#include "Common/determineDefaultDeviceNum.h"
 
 namespace qmcplusplus
 {
