+++ src/Platforms/CUDA/MemManageCUDA.hpp
@@ -24,10 +24,10 @@
 #include <stdexcept>
 #include <atomic>
 #include <limits>
-#include <Common/MemManage.hpp>
+#include "Common/MemManage.hpp"
+#include "Common/allocator_traits.hpp"
 #include "QueueCUDA.hpp"
 #include "CUDAruntime.hpp"
-#include "allocator_traits.hpp"
 #include "CUDAfill.hpp"
 
 namespace qmcplusplus
