+++ src/Platforms/Common/PlatformKinds.hpp
