+++ src/Platforms/Common/determineDefaultDeviceNum.h
